feat: move logout in profile settings
